% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoTransQF-package.R
\docType{package}
\name{AutoTransQF-package}
\alias{AutoTransQF-package}
\title{A Novel Automatic Shifted Log Transformation}
\description{
The R package AutoTransQF based on Feng et al.(2016) introduces a novel parametrization of log transformation and a shift parameter to automate the transformation process. Please read Feng et al. (2016) <doi: 10.1002/sta4.104> for more details of the method.
}
\author{
Yue Hu [aut, cre], Hyeon Lee [aut], J. S. Marron [aut]
}
