% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Melanoma}
\alias{Melanoma}
\title{MelanomaFeatures from Miedema et al. (2012)}
\format{
A data frame with 10152 observations on 49 variables where columns serve as data objects and rows serve as features. All 49 features are numeric variables including \code{Area}, \code{Hu.1}, \code{Hu.2 }etc.
}
\source{
Miedema, Jayson, et al. (2012). Image and statistical analysis of melanocytic histology. Histopathology, 61(3), pp.436-444. doi: 10.1111/j.1365-2559.2012.04229.x
}
\usage{
Melanoma
}
\description{
The dataset from Miedema et al. (2012) is built based on 49 hematoxylin and eosin stained slides of distinctive melanocytic lesions.
}
\keyword{datasets}
